refactor : 타입 원래대로 변경

+++ bank/src/main/java/com/bank/card_history/repository/CardHistoryRepositoryCustom.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface CardHistoryRepositoryCustom {
     List<CardHistory> getCardHistory(HistoryFindDto historyFindDto);
 
-    List<CategoryDto> getCardHistoryByCategory(List<Integer> userCardList, double totalPayByMonth, double totalDiscountByMonth, int month);
+    List<CategoryDto> getCardHistoryByCategory(List<Integer> userCardList, long totalPayByMonth, int totalDiscountByMonth, int month);
 
     TotalByMonth getTotalByMonth(List<Integer> cardUuidList, int month);
 }
